JCF: DUNE-DAQ/daq-deliverables#105: add a draft of the image one layer between the base image and the image with the Python packages
Note that this isn't a literal union of all the packages currently
getting yum-installed in the Dockerfiles in this repo. This is because
some packages are already available in the base image (e.g., bzip2 or
openssl), or in the case of the oracle client software, it doesn't
appear to be available in the repos reachable from this image. It _is_
available in the cern/cc7-base image, but the problem there is that
there's no Python version past 3.6 which can be installed by yum
(though to be fair, it could be installed manually if push came to
shove)