JCF: DUNE-DAQ/daq-deliverables#105: update base image in cern/alma9-base and add requirements.txt
For the time being will just use the Python which comes with cern/alma9-base
(v3.9.16), as this appears to be tied in to yum itself via the dnf Python package.
The Python packages in the requirements.txt are the union of the packages used
for the current microservices on the develop branch, minus cx-Oracle both because
Oracle use is uncertain moving forward and because it requires a dataclasses
version (0.8) which appears to be unavailable according to pip.

The versions of the Python packages are preferentially (1) the version found
in our nightly release, (2) the newest version across the microservice Python
packages, (3) in the case where no version was originally provided, the default
version which pip3 installed.